+++ packages/node-opcua-address-space/src/index_current.ts
@@ -11,7 +11,9 @@ export * from "../source/helpers/check_event_clause";
 export * from "../source/helpers/argument_list";
 export * from "../source/helpers/call_helpers";
 export * from "../source/helpers/ensure_secure_access";
+export * from "../source/helpers/resolve_opaque_on_address_space";
 export * from "../source/interfaces/alarms_and_conditions/condition_info_i";
+
 export { UACertificateExpirationAlarm } from "./alarms_and_conditions";
 
 export { ensureDatatypeExtracted, ensureDatatypeExtractedWithCallback } from "../source/loader/load_nodeset2";
