+++ packages/node-opcua-address-space/test/alarms_and_conditions/test_alarms_and_conditions.ts
@@ -2,8 +2,8 @@ import * as fs from "fs";
 import { nodesets } from "node-opcua-nodesets";
 import * as should from "should";
 
+import { generateAddressSpace } from "../../distNodeJS";
 import { AddressSpace } from "../..";
-import { generateAddressSpace } from "../../nodeJS";
 
 import { utest_acknowledgeable_condition } from "./utest_acknowledgeable_condition";
 import { utest_alarm_condition } from "./utest_alarm_condition";
