+++ packages/node-opcua-address-space/test/data_access/subtest_Y_array_item_type.ts
@@ -3,10 +3,10 @@ import * as should from "should";
 import { AxisInformation, AxisScaleEnumeration, standardUnits } from "node-opcua-data-access";
 import { coerceLocalizedText } from "node-opcua-data-model";
 import { resolveNodeId } from "node-opcua-nodeid";
-import { AxisInformationOptions } from "node-opcua-types";
 import { Variant } from "node-opcua-variant";
 import { DataType } from "node-opcua-variant";
 import { VariantArrayType } from "node-opcua-variant";
+
 import { AddressSpace, Folder, Namespace } from "../..";
 
 export function subtest_Y_array_item_type(maintest: any) {
