+++ packages/node-opcua-address-space/test/data_access/subtest_analog_item_semantic_changed.ts
@@ -7,6 +7,7 @@ import { DataValue } from "node-opcua-data-value";
 import { WriteValue } from "node-opcua-service-write";
 import { Variant } from "node-opcua-variant";
 import { DataType } from "node-opcua-variant";
+
 import { AddressSpace, SessionContext } from "../..";
 import { UAAnalogItem } from "../..";
