+++ packages/node-opcua-client-dynamic-extension-object/source/index.ts
@@ -1,7 +1,11 @@
 /**
  * @module node-opcua-client-dynamic-extension-object
  */
-export * from "./client_dynamic_extension_object";
+export * from "./populate_data_type_manager";
 export * from "./extra_data_type_manager";
 export * from "./resolve_dynamic_extension_object";
 export * from "./promote_opaque_structure";
+export * from "./get_extension_object_constructor";
+export * from "./get_extra_data_type_manager";
+export * from "./resolve_dynamic_extension_object";
+export * from "./convert_data_type_definition_to_structuretype_schema";
