+++ packages/node-opcua-client/source/private/client_session_impl.ts
@@ -7,12 +7,9 @@ import { assert } from "node-opcua-assert";
 import { AggregateFunction } from "node-opcua-constants";
 import { DateTime } from "node-opcua-basic-types";
 import {
-    extractNamespaceDataType,
     ExtraDataTypeManager,
-    getDataTypeDefinition,
     getExtensionObjectConstructor,
     getExtraDataTypeManager,
-    populateDataTypeManager,
     promoteOpaqueStructure,
     resolveDynamicExtensionObject
 } from "node-opcua-client-dynamic-extension-object";
@@ -20,7 +17,7 @@ import { Certificate, Nonce } from "node-opcua-crypto";
 import { attributeNameById, BrowseDirection, LocalizedTextLike } from "node-opcua-data-model";
 import { DataValue } from "node-opcua-data-value";
 import { checkDebugFlag, make_debugLog, make_errorLog, make_warningLog } from "node-opcua-debug";
-import { ExtensionObject, OpaqueStructure } from "node-opcua-extension-object";
+import { ExtensionObject } from "node-opcua-extension-object";
 import { coerceNodeId, NodeId, NodeIdLike, resolveNodeId } from "node-opcua-nodeid";
 import { getBuiltInDataType, getArgumentDefinitionHelper, IBasicSession } from "node-opcua-pseudo-session";
 import { AnyConstructorFunc } from "node-opcua-schemas";
@@ -95,13 +92,11 @@ import {
     BrowseNextRequest,
     BrowseNextResponse,
     HistoryReadValueIdOptions,
-    HistoryReadValueId,
     WriteValueOptions
 } from "node-opcua-types";
-import { buffer_ellipsis, check_flag, getFunctionParameterNames, isNullOrUndefined, lowerFirstLetter } from "node-opcua-utils";
+import { buffer_ellipsis, getFunctionParameterNames, isNullOrUndefined, lowerFirstLetter } from "node-opcua-utils";
 import { DataType, Variant, VariantLike } from "node-opcua-variant";
 
-import { DataTypeFactory, getStandardDataTypeFactory, StructuredTypeSchema } from "node-opcua-factory";
 import {
     ArgumentDefinition,
     BrowseDescriptionLike,
@@ -2041,7 +2036,7 @@ export class ClientSessionImpl extends EventEmitter implements ClientSession {
     ): void { }
 
     public async extractNamespaceDataType(): Promise<ExtraDataTypeManager> {
-        return extractNamespaceDataType(this);
+        return getExtraDataTypeManager(this);
     }
     public async getExtensionObjectConstructor(dataTypeNodeId: NodeId): Promise<AnyConstructorFunc> {
         return getExtensionObjectConstructor(this, dataTypeNodeId);
@@ -2168,7 +2163,7 @@ async function promoteOpaqueStructure2(session: IBasicSession, callMethodResult:
 
     const promises = callMethodResult.outputArguments.map(async (value: Variant) => {
         if (value.dataType === DataType.ExtensionObject) {
-            await resolveDynamicExtensionObject(value, extraDataTypeManager);
+            await resolveDynamicExtensionObject(session, value, extraDataTypeManager);
         }
     });
     await Promise.all(promises);
