+++ packages/node-opcua-modeler/tsconfig_nodejs.json
@@ -20,6 +20,6 @@
         }
     ],
     "exclude": [
-        "node_modules"
+        "node_modules" , "dist"
     ]
 }
