+++ packages/node-opcua-pseudo-session/source/basic_session_interface.ts
@@ -38,16 +38,17 @@ export interface IBasicSession {
      *
      * @param continuationPoint
      * @param releaseContinuationPoints  a Boolean parameter with the following values:
-     *      TRUE passed continuationPoints shall be reset to free resources in
+     *     * `true` passed continuationPoints shall be reset to free resources in
      *      the Server. The continuation points are released and the results
      *      and diagnosticInfos arrays are empty.
-     *      FALSE passed continuationPoints shall be used to get the next set of
+     *     * `false` passed continuationPoints shall be used to get the next set of
      *      browse information.
-     *      A Client shall always use the continuation point returned by a Browse or
-     *      BrowseNext response to free the resources for the continuation point in the
-     *      Server. If the Client does not want to get the next set of browse information,
-     *      BrowseNext shall be called with this parameter set to TRUE.
-     * @param callback
+     *     
+     *   A Client shall always use the continuation point returned by a Browse or
+     *    BrowseNext response to free the resources for the continuation point in the
+     *    Server. If the Client does not want to get the next set of browse information,
+     *    BrowseNext shall be called with this parameter set to `true`.
+     * 
      */
     browseNext(continuationPoint: Buffer, releaseContinuationPoints: boolean, callback: ResponseCallback<BrowseResult>): void;
 
@@ -104,9 +105,9 @@ export interface IBasicSession {
 
 export interface IBasicSession {
     write(nodeToWrite: WriteValueOptions, callback: ResponseCallback<StatusCode>): void;
-    
+
     write(nodesToWrite: WriteValueOptions[], callback: ResponseCallback<StatusCode[]>): void;
-    
+
     write(nodeToWrite: WriteValueOptions): Promise<StatusCode>;
 
     write(nodesToWrite: WriteValueOptions[]): Promise<StatusCode[]>;
@@ -196,15 +197,15 @@ export function getArgumentDefinitionHelper(
     });
 }
 
-export async function readNamespaceArray(session: IBasicSession): Promise<string[]>
-{
+export async function readNamespaceArray(session: IBasicSession): Promise<string[]> {
+    const nodeId = resolveNodeId(VariableIds.Server_NamespaceArray);
     const dataValue = await session.read({
-        nodeId: resolveNodeId(VariableIds.Server_NamespaceArray), 
+        nodeId,
         attributeId: AttributeIds.Value
     });
     if (dataValue.statusCode !== StatusCodes.Good) {
+        // errorLog("namespaceArray is not populated ! Your server must expose a list of namespaces in node ", nodeId.toString());
         return [];
     }
     return dataValue.value.value as string[];
 }
-
